update from source

+++ app/models/theme.rb
@@ -4,8 +4,6 @@ require_dependency 'stylesheet/manager'
 
 class Theme < ActiveRecord::Base
 
-  ALLOWED_FIELDS = %w{scss embedded_scss head_tag header after_header body_tag footer}
-
   @cache = DistributedCache.new('theme')
 
   belongs_to :color_scheme
@@ -46,6 +44,18 @@ class Theme < ActiveRecord::Base
     if SiteSetting.default_theme_key == self.key
       Theme.clear_default!
     end
+
+    if self.id
+
+      ColorScheme
+        .where(theme_id: self.id)
+        .where("id NOT IN (SELECT color_scheme_id FROM themes where color_scheme_id IS NOT NULL)")
+        .destroy_all
+
+      ColorScheme
+        .where(theme_id: self.id)
+        .update_all(theme_id: nil)
+    end
   end
 
   after_commit ->(theme) do
@@ -85,6 +95,10 @@ class Theme < ActiveRecord::Base
     Theme.expire_site_cache!
   end
 
+  def default?
+    SiteSetting.default_theme_key == key
+  end
+
   def self.lookup_field(key, target, field)
     return if key.blank?
 
@@ -100,7 +114,7 @@ class Theme < ActiveRecord::Base
     (@cache[cache_key] = val || "").html_safe
   end
 
-  def self.remove_from_cache!(themes=nil)
+  def self.remove_from_cache!(themes = nil)
     clear_cache!
   end
 
@@ -108,13 +122,11 @@ class Theme < ActiveRecord::Base
     @cache.clear
   end
 
-
   def self.targets
     @targets ||= Enum.new(common: 0, desktop: 1, mobile: 2)
   end
 
-
-  def notify_scheme_change(clear_manager_cache=true)
+  def notify_scheme_change(clear_manager_cache = true)
     Stylesheet::Manager.cache.clear if clear_manager_cache
     message = refresh_message_for_targets(["desktop", "mobile", "admin"], self)
     MessageBus.publish('/file-change', message)
@@ -126,7 +138,7 @@ class Theme < ActiveRecord::Base
     themes = [self] + dependant_themes
 
     message = themes.map do |theme|
-      refresh_message_for_targets([:mobile_theme,:desktop_theme], theme)
+      refresh_message_for_targets([:mobile_theme, :desktop_theme], theme)
     end.compact.flatten
     MessageBus.publish('/file-change', message)
   end
@@ -154,7 +166,7 @@ class Theme < ActiveRecord::Base
 
   def resolve_dependant_themes(direction)
 
-    select_field,where_field=nil
+    select_field, where_field = nil
 
     if direction == :up
       select_field = "parent_theme_id"
@@ -198,7 +210,7 @@ class Theme < ActiveRecord::Base
   end
 
   def resolve_baked_field(target, name)
-    list_baked_fields(target,name).map{|f| f.value_baked || f.value}.join("\n")
+    list_baked_fields(target, name).map { |f| f.value_baked || f.value }.join("\n")
   end
 
   def list_baked_fields(target, name)
@@ -206,13 +218,16 @@ class Theme < ActiveRecord::Base
     target = target.to_sym
 
     theme_ids = [self.id] + (included_themes.map(&:id) || [])
-    fields = ThemeField.where(target: [Theme.targets[target], Theme.targets[:common]])
-                       .where(name: name.to_s)
-                       .includes(:theme)
-                       .joins("JOIN (
-                             SELECT #{theme_ids.map.with_index{|id,idx| "#{id} AS theme_id, #{idx} AS sort_column"}.join(" UNION ALL SELECT ")}
-                            ) as X ON X.theme_id = theme_fields.theme_id")
-                       .order('sort_column, target')
+    fields = ThemeField.where(target_id: [Theme.targets[target], Theme.targets[:common]])
+      .where(name: name.to_s)
+      .includes(:theme)
+      .joins("
+        JOIN (
+          SELECT #{theme_ids.map.with_index { |id, idx| "#{id} AS theme_id, #{idx} AS sort_column" }.join(" UNION ALL SELECT ")}
+        ) as X ON X.theme_id = theme_fields.theme_id"
+      )
+      .order('sort_column, target_id')
+
     fields.each(&:ensure_baked!)
     fields
   end
@@ -229,24 +244,30 @@ class Theme < ActiveRecord::Base
     @changed_colors ||= []
   end
 
-  def set_field(target, name, value)
+  def set_field(target:, name:, value: nil, type: nil, type_id: nil, upload_id: nil)
     name = name.to_s
 
     target_id = Theme.targets[target.to_sym]
     raise "Unknown target #{target} passed to set field" unless target_id
 
-    field = theme_fields.find{|f| f.name==name && f.target == target_id}
+    type_id ||= type ? ThemeField.types[type.to_sym] : ThemeField.guess_type(name)
+    raise "Unknown type #{type} passed to set field" unless type_id
+
+    value ||= ""
+
+    field = theme_fields.find { |f| f.name == name && f.target_id == target_id && f.type_id == type_id }
     if field
-      if value.blank?
+      if value.blank? && !upload_id
         theme_fields.delete field.destroy
       else
-        if field.value != value
+        if field.value != value || field.upload_id != upload_id
           field.value = value
+          field.upload_id = upload_id
           changed_fields << field
         end
       end
     else
-      theme_fields.build(target: target_id, value: value, name: name) if value.present?
+      theme_fields.build(target_id: target_id, value: value, name: name, type_id: type_id, upload_id: upload_id) if value.present? || upload_id.present?
     end
   end
